pub_sub_app/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
google-cloud-pubsub==2.8.0